Refactor from extension method to own container
